14/6/21

+++ app/src/main/java/com/example/restaurant_app/modelmanager/gettingorder/Order.java
@@ -16,7 +16,7 @@ public class Order {
     private String paymentStatus;
     @SerializedName("grandTotal")
     @Expose
-    private Integer grandTotal;
+    private Double grandTotal;
     @SerializedName("paymentMethod")
     @Expose
     private String paymentMethod;
@@ -79,11 +79,11 @@ public class Order {
         this.paymentStatus = paymentStatus;
     }
 
-    public Integer getGrandTotal() {
+    public Double getGrandTotal() {
         return grandTotal;
     }
 
-    public void setGrandTotal(Integer grandTotal) {
+    public void setGrandTotal(Double grandTotal) {
         this.grandTotal = grandTotal;
     }
